[tsan] Refactor Memory Release Functions.
Summary:
This commit refactors calls to ReleaseMemoryPagesToOS into calls to
ReleaseTSANPages, and unmap/map to zero fill memory into ZeroPages.

This is in preparation for porting TSAN for Fuchsia: fuchsia doesn't
implement ReleaseMemoryPagesToOS, so we cannot call it directly.
Instead we will call an internal function that is able to only return
memory used for the tsan data structures (shadow, meta-shadow and
traces).

Reviewers: dvyukov

Subscribers: matthew, mcgrathr